[1501] Move mapping classes out of SubjectMapperService
Also rename StaticMapper => UCASSubjectToDFESubjectStaticMapper to make
its function more explicit.
