+++ config/initializers/inflections.rb
@@ -19,4 +19,6 @@ ActiveSupport::Inflector.inflections do |inflect|
   inflect.acronym 'API'
   inflect.acronym 'PGDE'
   inflect.acronym 'UCAS'
+  inflect.acronym 'MFL' # Modern foreign languages
+  inflect.acronym 'DFE' # Department for Educatino
 end
