+++ spec/services/subject_mapper_service_spec.rb
@@ -1,4 +1,3 @@
-require "spec_helper"
 require "csv"
 
 describe SubjectMapperService do
